split toolbar from document viewer main file
