+++ src/components/document/DocumentProvider.js
@@ -1,9 +1,8 @@
 import React, { createContext, useContext, useEffect, useState } from 'react'
-import { useSearch } from '../search/SearchProvider'
 import { useHashState } from '../HashStateProvider'
+import { TagsProvider } from './TagsProvider'
 import { collectionUrl, documentViewUrl } from '../../utils'
-import { createDownloadUrl, createTag, deleteTag, doc as docAPI, tags as tagsAPI, updateTag } from '../../backend/api'
-import { publicTagsList } from '../../constants/specialTags'
+import { createDownloadUrl, doc as docAPI } from '../../backend/api'
 
 const DocumentContext = createContext({})
 
@@ -22,8 +21,6 @@ export function DocumentProvider({ children, collection, collections, id, path, 
     const [tab, setTab] = useState(0)
     const [subTab, setSubTab] = useState(0)
     const { hashState, setHashState } = useHashState()
-
-    const { addTagToRefreshQueue } = useSearch()
 
     const collectionBaseUrl = collectionUrl(collection)
 
@@ -105,85 +102,6 @@ export function DocumentProvider({ children, collection, collections, id, path, 
         setHashState({ subTab: newValue }, false)
     }
 
-    const [tags, setTags] = useState([])
-    const [tagsLocked, setTagsLocked] = useState(false)
-    const [tagsLoading, setTagsLoading] = useState(true)
-    const [tagsError, setTagsError] = useState(null)
-
-    useEffect(() => {
-        if (digestUrl) {
-            setTagsLoading(true)
-            setTagsLocked(true)
-            setTagsError(null)
-            tagsAPI(digestUrl).then(data => {
-                setTags(data)
-                setTagsLoading(false)
-                setTagsLocked(false)
-            }).catch(res => {
-                setTags([])
-                setTagsError({ status: res.status, statusText: res.statusText, url: res.url })
-            }).finally(() => {
-                setTagsLoading(false)
-            })
-        }
-    }, [digestUrl])
-
-    const handleSpecialTagClick = (tag, name) => event => {
-        event.stopPropagation()
-        setTagsLocked(true)
-        if (tag) {
-            deleteTag(digestUrl, tag.id).then(() => {
-                setTags([...(tags.filter(t => t.id !== tag.id))])
-            }).finally(() => {
-                setTagsLocked(false)
-            })
-        } else {
-            createTag(digestUrl, { tag: name, public: publicTagsList.includes(name) }).then(newTag => {
-                setTags([...tags, newTag])
-                if (addTagToRefreshQueue) {
-                    addTagToRefreshQueue(digestUrl)
-                }
-            }).finally(() => {
-                setTagsLocked(false)
-            })
-        }
-    }
-
-    const handleTagAdd = (tag, publicTag = true) => {
-        setTagsLocked(true)
-        createTag(digestUrl, { tag, public: publicTagsList.includes(tag) || publicTag }).then(newTag => {
-            setTags([...tags, newTag])
-            if (addTagToRefreshQueue) {
-                addTagToRefreshQueue(digestUrl)
-            }
-        }).finally(() => {
-            setTagsLocked(false)
-        })
-    }
-
-    const handleTagDelete = tag => {
-        setTags([...tags])
-        setTagsLocked(true)
-        deleteTag(digestUrl, tag.id).then(() => {
-            setTags([...(tags.filter(t => t.id !== tag.id))])
-        }).catch(() => {
-            setTags([...tags])
-        }).finally(() => {
-            setTagsLocked(false)
-        })
-    }
-
-    const handleTagLockClick = tag => () => {
-        setTags([...tags])
-        setTagsLocked(true)
-        updateTag(digestUrl, tag.id, {public: !tag.public}).then(changedTag => {
-            Object.assign(tag, { ...changedTag })
-        }).finally(() => {
-            setTags([...tags])
-            setTagsLocked(false)
-        })
-    }
-
     return (
         <DocumentContext.Provider value={{
             data, pathname, loading, error,
@@ -192,11 +110,10 @@ export function DocumentProvider({ children, collection, collections, id, path, 
             digest, digestUrl, urlIsSha, docRawUrl,
             tab, handleTabChange,
             subTab, handleSubTabChange,
-            tags, tagsLocked, tagsLoading, tagsError,
-            handleSpecialTagClick, handleTagAdd,
-            handleTagDelete, handleTagLockClick,
         }}>
-            {children}
+            <TagsProvider>
+                {children}
+            </TagsProvider>
         </DocumentContext.Provider>
     )
 }
